refactor(mcp): split tools/card.rs into crud/relations/batch routers (KAN-762)
