+++ crates/kanban-mcp/src/tools/mod.rs
@@ -1,5 +1,7 @@
 pub(crate) mod board;
-pub(crate) mod card;
+pub(crate) mod card_batch;
+pub(crate) mod card_crud;
+pub(crate) mod card_relations;
 pub(crate) mod column;
 pub(crate) mod sprint;
 pub(crate) mod transfer;
